backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gspread==6.1.2
google-auth==2.35.0
anthropic==0.34.2
python-telegram-bot==21.6
apscheduler==3.10.4
cachetools==5.5.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-multipart==0.0.12
